bookmarks

+++ app/src/main/java/com/github/andreyasadchy/xtra/repository/OfflineRepository.kt
@@ -25,16 +25,10 @@ class OfflineRepository @Inject constructor(
         private val requestsDao: RequestsDao,
         private val localFollowsChannelDao: LocalFollowsChannelDao) {
 
-    fun loadAllVideosLiveData() = videosDao.getAllLiveData()
-
     fun loadAllVideos() = videosDao.getAll()
 
     suspend fun getVideoById(id: Int) = withContext(Dispatchers.IO) {
         videosDao.getById(id)
-    }
-
-    suspend fun getVideosByVideoId(videoId: String) = withContext(Dispatchers.IO) {
-        videosDao.getByVideoId(videoId)
     }
 
     suspend fun getVideosByUserId(id: Int) = withContext(Dispatchers.IO) {
@@ -48,13 +42,9 @@ class OfflineRepository @Inject constructor(
     fun deleteVideo(context: Context, video: OfflineVideo) {
         GlobalScope.launch {
             videosDao.delete(video)
-            if (video.channelId != null) {
-                if (videosDao.getByUserId(video.channelId.toInt()).isNullOrEmpty()) {
-                    File(context.filesDir.toString() + File.separator + "thumbnails" + File.separator + "${video.videoId}.png").delete()
-                }
-                if (localFollowsChannelDao.getById(video.channelId) == null) {
-                    File(context.filesDir.toString() + File.separator + "profile_pics" + File.separator + "${video.channelId}.png").delete()
-                }
+            File(context.filesDir.toString() + File.separator + "thumbnails" + File.separator + "${video.videoId}.png").delete()
+            if (video.channelId != null && localFollowsChannelDao.getById(video.channelId) == null) {
+                File(context.filesDir.toString() + File.separator + "profile_pics" + File.separator + "${video.channelId}.png").delete()
             }
         }
     }
